+++ src/jrd/ods.cpp
@@ -64,17 +64,17 @@ bool isSupported(USHORT majorVersion, USHORT minorVersion)
 	return false;
 }
 
-size_t bytesBitPIP(size_t page_size)
+ULONG bytesBitPIP(ULONG page_size)
 {
-	return page_size - OFFSETA(page_inv_page*, pip_bits);
+	return static_cast<ULONG>(page_size - OFFSETA(page_inv_page*, pip_bits));
 }
 
-size_t pagesPerPIP(size_t page_size)
+ULONG pagesPerPIP(ULONG page_size)
 {
 	return bytesBitPIP(page_size) * 8;
 }
 
-size_t pagesPerSCN(size_t page_size)
+ULONG pagesPerSCN(ULONG page_size)
 {
 	return pagesPerPIP(page_size) / BITS_PER_LONG;
 }
@@ -82,38 +82,38 @@ size_t pagesPerSCN(size_t page_size)
 // We must ensure that pagesPerSCN items can fit into scns_page::scn_pages array.
 // We can't use fb_assert() here in ods.h so it is placed at pag.cpp
 
-size_t maxPagesPerSCN(size_t page_size)
+ULONG maxPagesPerSCN(ULONG page_size)
 {
-	return (page_size - OFFSETA(scns_page*, scn_pages)) / sizeof(((scns_page*)NULL)->scn_pages);
+	return static_cast<ULONG>((page_size - OFFSETA(scns_page*, scn_pages)) / sizeof(((scns_page*)NULL)->scn_pages));
 }
 
-size_t transPerTIP(size_t page_size)
+ULONG transPerTIP(ULONG page_size)
 {
-	return (page_size - OFFSETA(tx_inv_page*, tip_transactions)) * 4;
+	return static_cast<ULONG>((page_size - OFFSETA(tx_inv_page*, tip_transactions)) * 4);
 }
 
-size_t gensPerPage(size_t page_size)
+ULONG gensPerPage(ULONG page_size)
 {
-	return (page_size - OFFSETA(generator_page*, gpg_values)) /
-		sizeof(((generator_page*) NULL)->gpg_values);
+	return static_cast<ULONG>((page_size - OFFSETA(generator_page*, gpg_values)) /
+		sizeof(((generator_page*) NULL)->gpg_values));
 }
 
-size_t dataPagesPerPP(size_t page_size)
+ULONG dataPagesPerPP(ULONG page_size)
 {
 	// Compute the number of data pages per pointer page.  Each data page
 	// requires a 32 bit pointer and a 4 bit control field.
 
-	return (page_size - OFFSETA(pointer_page*, ppg_page)) * 8 / (BITS_PER_LONG + PPG_DP_BITS_NUM);
+	return static_cast<ULONG>((page_size - OFFSETA(pointer_page*, ppg_page)) * 8 / (BITS_PER_LONG + PPG_DP_BITS_NUM));
 }
 
-size_t maxRecsPerDP(size_t page_size)
+ULONG maxRecsPerDP(ULONG page_size)
 {
 	// Compute the number of records that can fit on a page using the
 	// size of the record index (dpb_repeat) and a record header.  This
 	// gives an artificially high number, reducing the density of db_keys.
 
-	size_t max_records =
-		(page_size - sizeof(data_page)) / (sizeof(data_page::dpg_repeat) + OFFSETA(rhd*, rhd_data));
+	ULONG max_records =
+		static_cast<ULONG>((page_size - sizeof(data_page)) / (sizeof(data_page::dpg_repeat) + OFFSETA(rhd*, rhd_data)));
 
 	// Artificially reduce density of records to test high bits of record number
 	// max_records = 32000;
@@ -129,13 +129,13 @@ size_t maxRecsPerDP(size_t page_size)
 	return max_records;
 }
 
-size_t maxIndices(size_t page_size)
+ULONG maxIndices(ULONG page_size)
 {
 	// Compute the number of index roots that will fit on an index root page,
 	// assuming that each index has only one key
 
-	return (page_size - OFFSETA(index_root_page*, irt_rpt)) /
-		(sizeof(index_root_page::irt_repeat) + sizeof(irtd));
+	return static_cast<ULONG>((page_size - OFFSETA(index_root_page*, irt_rpt)) /
+		(sizeof(index_root_page::irt_repeat) + sizeof(irtd)));
 }
 
 Firebird::string pagtype(UCHAR type)
